pyproject.toml
[tool.poetry]
name = "my_blog"
version = "0.1.0"
description = ""
authors = ["ziwan"]

[tool.poetry.dependencies]
python = "3.10.5"
Django = "^4.0.6"
Pillow = "^9.2.0"
mysqlclient = "^2.1.1"
django-summernote = "0.8.20"
gunicorn = "^20.1.0"
django-on-heroku = "1.1.2"
prettyconf = "2.2.1"
psycopg2 = "^2.9.3"
ipdb = "^0.13.9"
imagekitio = "^2.2.8"
numpy = "^1.23.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
